project merge, multiple targets (library, tests, utility)
